Move getTime logic from MiddleCard to utils.
