+++ src/components/questionCard/questionCardBottom/QuestionCardBottom.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import styles from "./QuestionCardBottom.module.css";
 
 interface QuestionCardBottomProps {
@@ -11,7 +10,12 @@ export default function QuestionCardBottom({ data }: QuestionCardBottomProps) {
     return (
         <div className={styles.container}>
             {data.tags.map(tag => (
-                <p className={styles.tag_container}>{tag}</p>
+                <p
+                    key={tag}
+                    className={styles.tag_container}
+                >
+                    {tag}
+                </p>
             ))}
         </div>
     );
